[doc] rename SimpleNetwork in Network
