+++ Benchmarks/BenchmarkList.xml
@@ -7,9 +7,9 @@
                     <NetworkFixedSalinity>
                         <test> NetworkFixedSalinity_CI.xml </test>
                     </NetworkFixedSalinity>
-                    <SimpleNetwork>
-                        <test> SimpleNetwork_CI.xml </test>
-                    </SimpleNetwork>
+                    <Network>
+                        <test> Network_CI.xml </test>
+                    </Network>
                 </Belowground>
             </BettinaNetwork>
             <Bettina>
