+++ Benchmarks/ModuleBenchmarks/PlantModules/BettinaNetwork/Belowground/Network/Network.xml
@@ -5,7 +5,7 @@
             <type> SimpleTest </type>
         </aboveground>
         <belowground>
-            <type> SimpleNetwork </type>
+            <type> Network </type>
             <f_radius> 0.25 </f_radius>
         </belowground>
     </resources>
@@ -29,7 +29,7 @@
     <time_loop>
         <type> Simple </type>
         <t_start> 0 </t_start>
-        <t_end> 2e6 </t_end>
+        <t_end> 5e8 </t_end>
         <delta_t> 1e6 </delta_t>
     </time_loop>
     <visualization>
@@ -38,7 +38,7 @@
     <output>
         <type> OneTimestepOneFile </type>
         <output_each_nth_timestep>1</output_each_nth_timestep>
-        <output_times> [2e6] </output_times>
+        <output_times> [2e6, 5e8] </output_times>
         <allow_previous_output>True</allow_previous_output>
         <output_dir>Benchmarks/TestOutputs/</output_dir>
         <geometry_output> r_stem </geometry_output>
