+++ ResourceLib/BelowGround/Network/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-from .SimpleNetwork import SimpleNetwork
+from .Network import Network
 from .NetworkFixedSalinity import NetworkFixedSalinity
 from .NetworkOGSLargeScale3D import NetworkOGSLargeScale3D
